Removed Auth Logic From Routes
